reafactor: SearchBarViewController 삭제, 구조 단순화
- 네비게이션 바 구현 없이 검색결과를 나타내는 VC에 서치바를 추가했습니다.
