+++ ios5-Movie/ios5-Movie/SearchView/SearchListViewController.swift
@@ -8,10 +8,27 @@
 import UIKit
 import SnapKit
 
-class SearchListViewController: UIViewController {
+class SearchListViewController: UIViewController,UISearchBarDelegate {
     
     private let searchModel = SearchModel()
     private var movies: [Movie] = []
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        setupUI()
+    }
+    
+    //서치 컨트롤러 기본 설정
+    private lazy var searchController: UISearchController = {
+        //검색화면을 담당하는 VC
+        let searchController = UISearchController(searchResultsController: SearchListViewController())
+        searchController.searchBar.placeholder = "영화 이름 검색"
+        searchController.searchBar.delegate = self
+        searchController.hidesNavigationBarDuringPresentation = true
+        searchController.obscuresBackgroundDuringPresentation = false
+        
+        return searchController
+    }()
     
     private lazy var collectionView: UICollectionView = {
         let layout = UICollectionViewFlowLayout()
@@ -25,18 +42,6 @@ class SearchListViewController: UIViewController {
         return cv
     }()
     
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        setupUI()
-    }
-    
-    private func setupUI() {
-        view.addSubview(collectionView)
-        collectionView.snp.makeConstraints { make in
-            make.edges.equalToSuperview()
-        }
-    }
-    
     //검색결과 업데이트
     func updateSearchResults(with searchText: String?) {
         searchModel.updateSearchResults(with: searchText) { [weak self] movies in
@@ -45,6 +50,20 @@ class SearchListViewController: UIViewController {
                 self?.collectionView.reloadData()
             }
         }
+    }
+    
+    private func setupUI() {
+        //서치바 규약
+        [searchController.searchBar].forEach {
+            view.addSubview($0)}
+        //컬렉션뷰 규약
+        [collectionView].forEach { view.addSubview($0)}
+        collectionView.snp.makeConstraints {
+            $0.top.equalTo(searchController.searchBar.snp.bottom)
+            $0.leading.trailing.bottom.equalToSuperview()
+        }
+        //검색바가 화면을 덮지 않도록 설정
+        definesPresentationContext = true
     }
 }
 
@@ -67,5 +86,15 @@ extension SearchListViewController: UICollectionViewDataSource, UICollectionView
         let totalSpacing: CGFloat = (spacing * 2) + (spacing * 2)
         let width = (collectionView.bounds.width - totalSpacing) / 3
         return CGSize(width: width, height: width * 1.5)
+    }
+}
+
+extension SearchListViewController {
+    //서치바 입력값
+    func searchBar(_ searchBar: UISearchBar, textDidChange searchText: String) {
+        if let resultsController = searchController.searchResultsController as? SearchListViewController {
+            resultsController.updateSearchResults(with: searchText)
+        }
+        print("\(searchText)")
     }
 }
